Fix issues building with Visual Studio
Some missing files in CMake, a few compiler warnings.

+++ lib/gensio.c
@@ -2720,7 +2720,7 @@ gensio_get_default(struct gensio_os_funcs *o,
 		err = GE_NOMEM;
 		goto out_unlock;
 	    }
-	    memcpy(str, val->strval, val->intval + 1L); /* copy emding \0 */
+	    memcpy(str, val->strval, (size_t) val->intval + 1); /* copy emding \0 */
 	    *strval = str;
 	    *intval = val->intval;
 	} else {
@@ -3509,7 +3509,7 @@ gensio_alloc_vsprintf(struct gensio_os_funcs *o, const char *fmt, va_list va)
     char c[1], *str;
 
     va_copy(va2, va);
-    len = vsnprintf(c, 0, fmt, va) + 1L;
+    len = (size_t) vsnprintf(c, 0, fmt, va) + 1L;
     str = o->zalloc(o, len);
     if (str)
 	vsnprintf(str, len, fmt, va2);
